requirements.txt
absl==0.0
attr==0.3.2
graphviz==0.21
matplotlib==3.5.0
mpi4py==4.1.1
numpy==1.21.4
Pillow==12.0.0
pycryptodome==3.21.0
scipy==1.8.0
six==1.16.0
tensorflow==2.20.0
tensorflow_federated==0.87.0
tensorflow_privacy==0.9.0
tensorflow_probability==0.25.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.6
wandb==0.22.3
